requirements.txt
streamlit==1.26.0
pg8000==1.29.6
python-dotenv==1.0.0